requirements.txt
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0